requirements.txt
APScheduler==3.10.1
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-APScheduler==1.12.4
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.2
tzlocal==5.0.1
Werkzeug==2.3.4
WTForms==3.0.1
